batches: Don't error if unknown event received
As GitHub evolves, they are adding more events. Also, we don't care about all possible ones, so it might well be that we don't know of some of the kinds of events we receive. Before, we'd return an error, which causes GitHub to retry the webhook eventually. That caused a lot of additional failures and incoming requests, additional load on the GitHub instance, and a lot of logspam. This fixes it by accepting that we can't know about _all_ events.